Mudanças básicas e implementação do background
