+++ Wake-Move/app-example/(tabs)/explore.tsx
